Handle more NFT display edge cases
